Frontend: Extract a component for displaying highstate
